+++ management/src/main/java/com/school/management/model/SessionYear.java
@@ -10,7 +10,7 @@ public class SessionYear {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     private int startingYear;
     private int endYear;
@@ -19,9 +19,9 @@ public class SessionYear {
     @JoinColumn(name = "session_year_id")
     private List<SClass> classes;
 
-    public long getId() { return id; }
+    public int getId() { return id; }
 
-    public void setId(long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
 
     public int getStartingYear() {
         return startingYear;
